Move scroll logic into popup module
